refactor!: move extension methods into container objects
